update class diagrams renamed visualise package to visualisation
