+++ VisualiseClasses.puml
@@ -1,32 +1,32 @@
 @startuml
 
+!include ClassDefs.iuml
+
 header Scott Logic DataHelix Project
 footer 15 February 2019
 title Visualisation Class Hierarchy
 
-abstract CommandLineBase
-abstract Runnable
-abstract CommandLineBase
-
-interface ConfigSource
-interface VisualisationConfigSource
-
 ConfigSource <|-- VisualisationConfigSource
-CommandLineBase <|-- VisualisationCommandLine
+CommandLineBase <|-- VisualiseCommandLine
 
 CommandLineBase --|> ConfigSource
 Runnable <|-- CommandLineBase
 
-VisualiseConfigValidator --* OutputTarget
-VisualiseConfigValidator --* VisualisationConfigSource
+VisualisationConfigValidator --* OutputTarget
+VisualisationConfigValidator --* VisualisationConfigSource
 
 VisualiseExecute --|> Runnable
 VisualiseExecute --* VisualisationConfig
 VisualiseExecute --* ProfileReader
 VisualiseExecute --* VisualisationConfigSource
-VisualiseExecute --* VisualiseConfigValidator
+VisualiseExecute --* VisualisationConfigValidator
 VisualiseExecute --* ErrorReporter
 
-OutputTarget <|-- FileOutputTarget
+
+hide GenerateExecute
+hide GenerationConfig
+hide GenerationConfigSource
+hide GenerationConfigValidator
+hide StaticContradictionDecisionTreeValidator
 
 @enduml
