+++ generator/src/main/java/com/scottlogic/deg/generator/visualisation/VisualisationConfig.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.visualise;
+package com.scottlogic.deg.generator.visualisation;
 
 import com.google.inject.Inject;
 import com.scottlogic.deg.generator.inputs.validation.NoopProfileValidator;
