+++ generator/src/test/java/com/scottlogic/deg/generator/visualisation/TestVisualiseConfigSource.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.visualise;
+package com.scottlogic.deg.generator.visualisation;
 
 import java.io.File;
 import java.nio.file.Path;
